import-graphql-files-use-stop-places-with-departures * Move useStopPlacesWithDepartures to its own directory * Extract graphql query to its own file and add @api directive * Change from using old apolloClient to new
