+++ src/logic/index.ts
@@ -1,6 +1,6 @@
 export { useBikeRentalStations } from './useBikeRentalStations'
 export { useNearestPlaces } from './useNearestPlaces'
-export { useStopPlacesWithDepartures } from './useStopPlacesWithDepartures'
+export { useStopPlacesWithDepartures } from './use-stop-places-with-departures/useStopPlacesWithDepartures'
 export { useMobility } from './useMobility'
 export { useWalkInfo } from './useWalkInfo'
 export { useWeather } from './useWeather'
