+++ src/logic/use-stop-places-with-departures/GetStopPlacesWithDepartures.graphql
@@ -1,0 +1,45 @@
+query getStopPlacesWithDepartures($ids: [String]!) @api(name: journey_planner_v3) {
+    stopPlaces(ids: $ids) {
+        id
+        name
+        description
+        latitude
+        longitude
+        transportMode
+        transportSubmode
+        estimatedCalls(
+            numberOfDepartures: 200
+            timeRange: 172800
+            numberOfDeparturesPerLineAndDestinationDisplay: 20
+            arrivalDeparture: departures
+        ) {
+            aimedDepartureTime
+            cancellation
+            date
+            destinationDisplay {
+                frontText
+            }
+            expectedDepartureTime
+            quay {
+                id
+                name
+                publicCode
+            }
+            serviceJourney {
+                id
+                journeyPattern {
+                    line {
+                        publicCode
+                        transportMode
+                    }
+                }
+                transportSubmode
+            }
+            situations {
+                summary {
+                    value
+                }
+            }
+        }
+    }
+}
